리뷰 평점기능 구현

+++ src/main/java/com/imageupload/example/entity/ProfileEntity.java
@@ -21,7 +21,7 @@ public class ProfileEntity implements Serializable {
 
     @Id @GeneratedValue
     private Long id;
-    private int mannerScore;
+    private float mannerScore;
     private int mileage;
     private String nickname;
     private String profilePath;
